refactor: Split SkinCitizen into partials
